Cart properties i.e. total are now BigDecimals not Money.
- Cart should only care about 77.50 not £77.50. Maybe in future we'll have a reference to the local
on the cart so elsewhere we can go `Money.new(cart.total, cart.local)`

+++ app/lib/modifiers.rb
@@ -11,8 +11,7 @@ module Modifiers
 
   class PercentageModifier < Modifier
     def initialize(percentage)
-      # TODO do floats work? Check. Make sure we don't get decimal BUGS
-      @percentage = percentage.to_f/100.to_f
+      @percentage = percentage.to_d/100.to_d
     end
   end
